+++ CHANGELOG.md
@@ -1,3 +1,80 @@
+# [1.325.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.324.1...v1.325.0) (2026-02-25)
+
+
+### Features
+
+* LLMO-3163: byocdn-other scans and triggers audits for any day in… ([#2020](https://github.com/adobe/spacecat-audit-worker/issues/2020)) ([2c83095](https://github.com/adobe/spacecat-audit-worker/commit/2c83095a82372da77b4781854c854d7fe6d0aa57))
+
+## [1.324.1](https://github.com/adobe/spacecat-audit-worker/compare/v1.324.0...v1.324.1) (2026-02-25)
+
+
+### Bug Fixes
+
+* Filter out non-actionable redundant permissions ([#1967](https://github.com/adobe/spacecat-audit-worker/issues/1967)) ([34fbc78](https://github.com/adobe/spacecat-audit-worker/commit/34fbc7840f16b7ec065f8619d46185010a887966))
+
+# [1.324.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.323.0...v1.324.0) (2026-02-25)
+
+
+### Features
+
+* migrate secrets from AWS SM to Vault ([#2033](https://github.com/adobe/spacecat-audit-worker/issues/2033)) ([6072d11](https://github.com/adobe/spacecat-audit-worker/commit/6072d11407f0d0d0ac215b5e76a50f2bcab6f181))
+
+# [1.323.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.322.7...v1.323.0) (2026-02-24)
+
+
+### Features
+
+* Sitemap audit provides tuning for bot protection ([#2029](https://github.com/adobe/spacecat-audit-worker/issues/2029)) ([c782040](https://github.com/adobe/spacecat-audit-worker/commit/c78204060df0135b94d4ce5b6bb8a796011948cc))
+
+## [1.322.7](https://github.com/adobe/spacecat-audit-worker/compare/v1.322.6...v1.322.7) (2026-02-24)
+
+
+### Bug Fixes
+
+* update existing opportunity instead of marking as ignored ([#2030](https://github.com/adobe/spacecat-audit-worker/issues/2030)) ([d95f2f3](https://github.com/adobe/spacecat-audit-worker/commit/d95f2f39ea01dfbcf5fc5593d88e0ba475fa33b7))
+
+## [1.322.6](https://github.com/adobe/spacecat-audit-worker/compare/v1.322.5...v1.322.6) (2026-02-24)
+
+
+### Bug Fixes
+
+* zero-pad week number in llm-error-pages filenames ([#2024](https://github.com/adobe/spacecat-audit-worker/issues/2024)) ([5662f23](https://github.com/adobe/spacecat-audit-worker/commit/5662f237e7c6a4499a0656872ddad7f4f25a5b46))
+
+## [1.322.5](https://github.com/adobe/spacecat-audit-worker/compare/v1.322.4...v1.322.5) (2026-02-24)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1768](https://github.com/adobe/spacecat-audit-worker/issues/1768)) ([76e8675](https://github.com/adobe/spacecat-audit-worker/commit/76e86754a7c2416e55c76a4c326b4aba679aeffd))
+
+## [1.322.4](https://github.com/adobe/spacecat-audit-worker/compare/v1.322.3...v1.322.4) (2026-02-24)
+
+
+### Bug Fixes
+
+* [canonical] validate canonical URL accessibility for different domain or protocol ([#2014](https://github.com/adobe/spacecat-audit-worker/issues/2014)) ([20ae30a](https://github.com/adobe/spacecat-audit-worker/commit/20ae30a9a85fca7f948baf928882e18f7eb585e0))
+
+## [1.322.3](https://github.com/adobe/spacecat-audit-worker/compare/v1.322.2...v1.322.3) (2026-02-23)
+
+
+### Bug Fixes
+
+* avoid generating guidances for excluded forms ([#2017](https://github.com/adobe/spacecat-audit-worker/issues/2017)) ([7c960af](https://github.com/adobe/spacecat-audit-worker/commit/7c960af6cc382df7b5f92ab2a133dd174cf00e47))
+
+## [1.322.2](https://github.com/adobe/spacecat-audit-worker/compare/v1.322.1...v1.322.2) (2026-02-23)
+
+
+### Bug Fixes
+
+* adding order form in ignored form lists for form opportunities ([#2021](https://github.com/adobe/spacecat-audit-worker/issues/2021)) ([a51df4f](https://github.com/adobe/spacecat-audit-worker/commit/a51df4f9638b77d52db8062a9efd015177a4b15c))
+
+## [1.322.1](https://github.com/adobe/spacecat-audit-worker/compare/v1.322.0...v1.322.1) (2026-02-19)
+
+
+### Bug Fixes
+
+* bright data suggestions rejected if locale does not match ([#2004](https://github.com/adobe/spacecat-audit-worker/issues/2004)) ([5486a34](https://github.com/adobe/spacecat-audit-worker/commit/5486a341ac02151fe4f6af41903cb3104f47a030))
+
 # [1.322.0](https://github.com/adobe/spacecat-audit-worker/compare/v1.321.1...v1.322.0) (2026-02-19)